feat: add turret template
Add a turret template which can shoot on enemies.
Add some entity statistic resource allowing to set stats to entities,
those stats are values like ap, armor, range, fire rate or damage.
Add a shader to change the color of an sprite via a matching table.
Add the possiblity to modify the stats of a turret.

Fix $14c