feat: add spawn generator
Add a spawn generator, this piece of code will try to spawn units until
a target strenght is reached.
Add a generator to calculate the target strengt per wave, this is done
by an exponential function.
Add possibility to add multiple spawners, this allows enemies to attack
from multiple sides.
Add method to enemies object to calculate there power, this is used to
define how many units need to be spawned per wave.