feat: start adding popup windows for tower
Add a popup to towers if they are getting clicked.

This popup will allow users to set the threat mode of the tower as well
as buy modifications each wave. Those modifications will increase the power of each of the towers.

Rework the modifier system by moving more information into the base
class.
Add modifier generator instance, which is used to generate modifiers
dynamically.

Fix $15w