feat: add player controlled units
Add a selection mode allowing the player to select units.
Add a unit template to allow the player to command units.
Add a selection sprite to make it more visible which units are selected
right now.
Refactored enemy script to be better reusable.
Make walk animations a resource object for better reuse.